Improve lock-free queue performance
Instead of using two variable `readable` and `writable` to create
a critical section, this patch implemented multiple consumer
single producer lock-free queue by using a variable `count` to
record number of available items in queue. One of the advantages of
that is the ability to implement atomic enqueues with relative ease.